Setup for React, JSX, ES6+, ESLint and shortcuts
For ESLint to work properly it’s recommended to 

`npm install -g babel-eslint eslint eslint-config-airbnb eslint-plugin-react`

You’ll need this if working on the web app and want to get anything
through CI.

It’d be rad to have Syntastic utilize the local packages within 
node_modules but it doesn’t really seem to be a thing

There’s this issue:
- https://github.com/scrooloose/syntastic/issues/1571

Then this plugin exists:
https://github.com/mtscout6/syntastic-local-eslint.vim

I’ll most likely play with it at some point. It’s really not too bad
keeping the global installs in sync with the web app.

We ignore console log errors but only in Vim. Usually CI or running
`npm test` is going to fail those.

And finally added a shortcut the same as our CoffeeScript setup.

`cdl` completes to `console.log(I)` where `I` is the cursor.