Remove Meslo and rock Apple’s Menlo font
Within terminal you can now customize the line height (which is in the
polarized.terminal file) so there is no need to download the custom font 
or even have it on your machine any longer. The only reason we were 
using Meslo was to get the taller line-height. I believe the glyphs are
the same. Keep it simple.