Adds the vim-flow plugin for running flowtype
Added a few settings:

1. Use the local version of flow in node_modules if it exists
2. Close Flow's QF window when there are no errors
3. Don't use vim-flow's QF sizing (it jacks with other QF settings)
4. Enable omni-completion by default (haven't played with this yet)

There is a syntastic version of this.. but it's sort of whacky. When
there is an error, flowtype reports 2 errors. One on the definition and
one on the caller. The syntastic version only links to the definition
and not the caller which is really probably the one you want. Also, the way
that flowtype works is a server is running in the background which
checks all files. With Syntastic, this means that errors are reported in
every file in their own individual QF windows. This plugin, while not
perfect gives a little better feedback and only in a single window.